requirements.txt
PyJWT==1.7.1
virtualenv==20.0.35
pip==20.2.4
dacite==1.5.1
pytest==6.1.1
Django==3.1.2
djangorestframework==3.11.0
dataclasses==0.6
typing==3.7.4.3
